check if files exist

+++ delete.py
@@ -1,4 +1,5 @@
 import os
+import pathlib
 # from main import base
 
 
@@ -7,4 +8,6 @@ def delete_directory(root, location):
 
 
 def delete_file(root, location):
+    if not pathlib.Path(root/location).exists():
+        exit('File does not exists')
     return os.remove(root/location)
